交易市场逻辑

+++ database/migrations/2020_10_24_150539_create_coupon_markets_table.php
@@ -18,7 +18,7 @@ class CreateCouponMarketsTable extends Migration
       $table->unsignedBigInteger('sell_user_id')->comment('出售人');
       $table->unsignedBigInteger('buy_user_id')->comment('购买人')->nullable();
       $table->unsignedBigInteger('user_coupon_id')->comment('优惠券id');
-      $table->unsignedInteger('coupon_template_id')->comment('优惠券模板id');
+      $table->unsignedBigInteger('coupon_template_id')->comment('优惠券模板id');
       $table->decimal('amount', 6, 2)->comment('出售价格');
       $table->unsignedInteger('amount_sort')->comment('按金额排序');
       $table->unsignedTinyInteger('status')->comment('出售状态0出售中1待支付2已出售3已下架4已撤回')->default(0);
